+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
 - API event extrude_override_change
 - API method set_speed_override
 - API event speed_override_change
+- API method set_fan_speed
+- API method fan_speed_change
 
 ## [0.14.0] - 2016-03-30
 ## Added